Add remove book and toggle read or not
displayBooksWithMap function added to:
Remove book and toggle read or not buttons given ids including
myLibrary index of the book.
Event listeners then added using the id, apply the remove and toggle
to the relevant book in myLibrary.
On remove or toggle, a new call of displayBooksWithMap is used
to refresh the list on the page.